fixzzzzz

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: brpereir <brpereir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/12 20:34:19 by brpereir          #+#    #+#             */
-/*   Updated: 2023/04/12 22:59:57 by brpereir         ###   ########.fr       */
+/*   Updated: 2023/04/14 18:21:43 by brpereir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void *calloc(size_t n, size_t size)
 {
-	int	i;
+	size_t	i;
 	char *arr;
 
 	i = 0;
